[#179218883] new errors
