+++ lib/sendcloud/errors/bad_request_error.rb
@@ -1,20 +1,5 @@
 module Sendcloud
   module Errors
-    class BadRequestError < StandardError
-      attr_reader :body, :response
-
-      def initialize(response:, body:)
-        @response = response
-        @body = body
-
-        super(build_message)
-      end
-
-      private
-
-      def build_message
-        "#{response.status} #{body[:status]} #{body[:message]}"
-      end
-    end
+    class BadRequestError < ClientError; end
   end
 end
