+++ lib/sendcloud/errors/deleted_error.rb
@@ -1,0 +1,5 @@
+module Sendcloud
+  module Errors
+    class DeletedError < ClientError; end
+  end
+end
